refactor: extract openai adapter tests
